# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SuuqLink marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendor storefronts
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: featured
          schema: { type: boolean }
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /vendors/{id}/products:
    get:
      operationId: getVendorProducts
      tags: [vendors]
      summary: List products for a vendor
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List/search products with filters
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
          description: Category slug
        - in: query
          name: vendorId
          schema: { type: integer }
        - in: query
          name: minPrice
          schema: { type: number }
        - in: query
          name: maxPrice
          schema: { type: number }
        - in: query
          name: sort
          schema:
            type: string
            enum: [popular, newest, price_asc, price_desc, rating]
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /search/suggestions:
    get:
      operationId: getSearchSuggestions
      tags: [products]
      summary: Get search suggestions
      parameters:
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchSuggestions"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List my orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Category:
      type: object
      required: [id, slug, name, icon, productCount, accentColor]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        icon: { type: string }
        accentColor: { type: string }
        productCount: { type: integer }
    Vendor:
      type: object
      required: [id, slug, name, tagline, city, country, rating, reviewCount, productCount, logoUrl, coverUrl, verified, badges]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        tagline: { type: string }
        city: { type: string }
        country: { type: string }
        rating: { type: number }
        reviewCount: { type: integer }
        productCount: { type: integer }
        logoUrl: { type: string }
        coverUrl: { type: string }
        verified: { type: boolean }
        badges:
          type: array
          items: { type: string }
    VendorDetail:
      allOf:
        - $ref: "#/components/schemas/Vendor"
        - type: object
          required: [about, joinedAt, responseTimeMins, fulfillmentRate, products]
          properties:
            about: { type: string }
            joinedAt: { type: string, format: date-time }
            responseTimeMins: { type: integer }
            fulfillmentRate: { type: number }
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
    Product:
      type: object
      required:
        - id
        - slug
        - title
        - price
        - currency
        - imageUrl
        - vendorId
        - vendorName
        - rating
        - reviewCount
        - categorySlug
        - categoryName
        - inStock
        - badges
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        price: { type: number }
        compareAtPrice:
          type: ["number", "null"]
        currency: { type: string }
        imageUrl: { type: string }
        vendorId: { type: integer }
        vendorName: { type: string }
        rating: { type: number }
        reviewCount: { type: integer }
        categorySlug: { type: string }
        categoryName: { type: string }
        inStock: { type: boolean }
        badges:
          type: array
          items: { type: string }
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          required: [description, gallery, specs, shippingDays, vendor, reviews]
          properties:
            description: { type: string }
            gallery:
              type: array
              items: { type: string }
            specs:
              type: array
              items:
                $ref: "#/components/schemas/ProductSpec"
            shippingDays: { type: integer }
            vendor:
              $ref: "#/components/schemas/Vendor"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
    ProductSpec:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }
    Review:
      type: object
      required: [id, author, rating, body, createdAt]
      properties:
        id: { type: integer }
        author: { type: string }
        rating: { type: number }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    ProductListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }
    SearchSuggestions:
      type: object
      required: [terms, categories, products]
      properties:
        terms:
          type: array
          items: { type: string }
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
...
    OrderItem:
      type: object
      required: [id, productId, title, imageUrl, vendorName, price, currency, quantity]
      properties:
        id: { type: integer }
        productId: { type: integer }
        title: { type: string }
        imageUrl: { type: string }
        vendorName: { type: string }
        price: { type: number }
        currency: { type: string }
        quantity: { type: integer }
    OrderTimelineEvent:
      type: object
      required: [status, label, at, completed]
      properties:
        status: { type: string }
        label: { type: string }
        at: { type: string, format: date-time }
        completed: { type: boolean }
    Order:
      type: object
      required:
        - id
        - reference
        - status
        - items
        - subtotal
        - shipping
        - tax
        - total
        - currency
        - shippingAddress
        - paymentMethod
        - createdAt
        - estimatedDelivery
        - timeline
      properties:
        id: { type: integer }
        reference: { type: string }
        status:
          type: string
          enum: [placed, confirmed, packed, shipped, out_for_delivery, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal: { type: number }
        shipping: { type: number }
        tax: { type: number }
        total: { type: number }
        currency: { type: string }
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        paymentMethod: { type: string }
        createdAt: { type: string, format: date-time }
        estimatedDelivery: { type: string, format: date-time }
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/OrderTimelineEvent"
    ShippingAddress:
      type: object
      required: [fullName, phone, line1, city, region, country]
      properties:
        fullName: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2:
          type: ["string", "null"]
        city: { type: string }
        region: { type: string }
        country: { type: string }
        notes:
          type: ["string", "null"]
    CreateOrderBody:
      type: object
      required: [shippingAddress, paymentMethod]
      properties:
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        paymentMethod:
          type: string
          enum: [mobile_money, card, cash_on_delivery, bank_transfer]
...