# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SuuqLink marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";
import type { OrderTimelineEvent } from "./orderTimelineEvent";
import type { ShippingAddress } from "./shippingAddress";

export interface Order {
  id: number;
  reference: string;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  tax: number;
  total: number;
  currency: string;
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  createdAt: Date;
  estimatedDelivery: Date;
  timeline: OrderTimelineEvent[];
}